Uploaded initial working files
latinwords.py generates the latinwords.txt dictionary from the website getting rid of junk in it and trying to structure the dictionary properly
PassGen.py is working right now prints two passwords, one is from the English dictionary which is from an MIT website list of words, 14 character random generated password made from two random words first one 6 characters long and the second one 5 characters long. The first letter of the first word is capitalized. Then joined with a random symbol chosen from only !#. attached at the beginning of the password and two random numbers joined to the end.
testing.py is the code to pull from the MIT website the two random words of a given character length.